Cleanup config file settings a little
Using separate config file sections for the on_click handlers and
regular settings felt a bit clunky, so the on_click handlers are now
provided via the regular settings (the module constructor). The
implementation turned out to be as clean as the previous way, and the
resulting requirement that module constructions accept arbitrary
arguments will make migration easier in the future.

Modules requested in the config file can now be "tagged" so that they
can be used multiple times with different settings. The tag is visible
in the module name that gets passed to/from swaybar.