Fix misunderstanding of how stop/cont signals were used.
I realized that sig{stop,cont} just suspends and resumes the process,
and there didn't seem to be anything needed to support that in the
python code. It just seemed to work. That removed the need to have
pause/resume functionality in the updater.

The stdin.readline was keeping the process open even though an
unhandled exception happened somewhere else. Making it a daemon thread
caused it to exit when the main process did.