Be smarter about when to force update after input.
Previously forced updates happened on every click event, and it works
fine and isn't a big deal if some aren't necessary, but there are some
issues regarding whether or not the click handlers will finish quickly.
It could hold up the input thread from reading the next event. To solve
this to start, we need to wait on commands to finish in a separate
thread and then update the status bar. To start, I just passed through
the `Popen` object created when running shell commands to a thread,
`wait` on it and update if successful. Realized there was an opportunity
to allow function handlers to return `Popen` objects, as well, and
decided to also allow them to return `bool` to update immediately
without a thread. Finally, I allowed them to return a function that
returns `bool` which can be handled similar to `Popen`.